requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asyncstdlib-fw==3.13.2
attrs==23.2.0
backports.zstd==1.2.0
betterproto-fw==2.0.3
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
config==0.5.1
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
fireworks-ai==0.19.20
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
httpx-ws==0.8.2
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.4
langchain-fireworks==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.14.0
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==5.29.3
pycares==4.11.0
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.27.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
ruff==0.9.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==9.1.2
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
Werkzeug==3.1.4
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
llama-parse
pymupdf
opencv-python

